+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## Version 20251127
+- Added default page option (and settings admin page)
+
 ## Version 20251126
 - Reorganized soruce code into feature-based architecture for better maintainability.
 - Only show changelog to admin users